# docs/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SME Digitization API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stores": {
      "post": {
        "tags": [
          "stores"
        ],
        "summary": "Create Store",
        "operationId": "create_store_api_v1_stores_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/me": {
      "get": {
        "tags": [
          "stores"
        ],
        "summary": "Get My Store",
        "operationId": "get_my_store_api_v1_stores_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/stores/{store_id}": {
      "patch": {
        "tags": [
          "stores"
        ],
        "summary": "Update Store",
        "operationId": "update_store_api_v1_stores__store_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sync/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Push",
        "operationId": "sync_push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/pull": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Pull",
        "operationId": "sync_pull_api_v1_sync_pull_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/status": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Status",
        "operationId": "sync_status_api_v1_sync_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoreCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locale_default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale Default"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "NPR"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StoreCreate"
      },
      "StoreOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locale_default": {
            "type": "string",
            "title": "Locale Default"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_user_id",
          "name",
          "locale_default",
          "currency",
          "created_at"
        ],
        "title": "StoreOut"
      },
      "StoreUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "locale_default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale Default"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          }
        },
        "type": "object",
        "title": "StoreUpdate"
      },
...
      "SyncPullEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "entity": {
            "type": "string",
            "title": "Entity"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity",
          "operation",
          "payload",
          "created_at"
        ],
        "title": "SyncPullEvent"
      },
      "SyncPullResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/SyncPullEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "SyncPullResponse"
      },
      "SyncPushEvent": {
        "properties": {
          "entity": {
            "type": "string",
            "title": "Entity"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "entity",
          "operation",
          "payload"
        ],
        "title": "SyncPushEvent"
      },
      "SyncPushRequest": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/SyncPushEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "SyncPushRequest"
      },
      "SyncStatusResponse": {
        "properties": {
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "last_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Id"
          },
          "recommended_pull_since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Pull Since"
          }
        },
        "type": "object",
        "required": [
          "server_time",
          "last_event_id",
          "recommended_pull_since"
        ],
        "title": "SyncStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}